% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relaxinfo.R
\name{relaxinfo}
\alias{relaxinfo}
\title{Display license information about embedded code}
\usage{
relaxinfo()
}
\value{
Some information about licenses of code and algorithms.
}
\description{
Function to display license information regarding code embedded in
\code{rrelaxiv}.
}
\examples{
relaxinfo()
}
