% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relaxiv.R
\name{.RELAX_IV}
\alias{.RELAX_IV}
\title{Lower-level access to the RELAX-IV algorithm}
\usage{
.RELAX_IV(n1, na1, startn1, endn1, c1, u1, b1, rc1, crash1, large1)
}
\arguments{
\item{n1}{Number of nodes}

\item{na1}{Number of arcs}

\item{startn1}{Starting nodes}

\item{endn1}{Ending nodes}

\item{c1}{Costs vector}

\item{u1}{Capacities}

\item{b1}{Demands}

\item{rc1}{Reduced Costs}

\item{crash1}{Set crash equal to 1 to activate an auction/shortest path
subroutine for getting the initial price-flow pair. this is recommended
for difficult problems where the default initialization yields long
solution times.}

\item{large1}{A very large integer to represent infinity}
}
\value{
A list containing all parameters, as well as `x1` which contains the
  solution, and `feasible1` which contains a flag for feasibility.
}
\description{
\code{.RELAX_IV} provides low-level access to the FORTRAN; taking the exact
input as the FORTRAN code. It does NOT do any input checking.
}
\seealso{
\code{\link{RELAX_IV}} for a more user-friendly interface.
}
