% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relaxiv.R
\name{RELAX_IV}
\alias{RELAX_IV}
\title{RELAX-IV Algorithm}
\usage{
RELAX_IV(startnodes, endnodes, arccosts, arccapacity, supply)
}
\arguments{
\item{startnodes}{A vector of starting nodes.}

\item{endnodes}{A vector of ending nodes.}

\item{arccosts}{A vector of arc costs.}

\item{arccapacity}{A vector of arc capacities.}

\item{supply}{A vector of supplies.}
}
\value{
The flow solution.
}
\description{
This provides an interface to the RELAX_IV minimum cost flow solver. It
provides input santization and returns the flow solution.
}
\examples{
data(relaxexample)
RELAX_IV(relaxexample$startnodes,
         relaxexample$endnodes,
         relaxexample$arccosts,
         relaxexample$arccapacity,
         relaxexample$nodedemand)
}
\seealso{
[.RELAX_IV()] for a more advanced interface and set of returned
  results.
}
