% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{relaxexample}
\alias{relaxexample}
\title{An example of a small network flow input.}
\format{
An object of class \code{list} of length 5.
}
\usage{
relaxexample
}
\description{
An example of a small network flow input.
}
\keyword{datasets}
